respond to pull request comments
Remove unneeded line about PUID and PGID. I had it in there because this solves a problem that the 'official' theLounge container has problems with. But I agree that the existing documentation of this feature is sufficient.

I admit I missed the markup for the URL in my preview. Now fixed and checked by pasting into README.md as a test edit.